int to uint

+++ comms.go
@@ -12,7 +12,7 @@ import (
 	"github.com/hink/go-blink1/libusb"
 )
 
-func fadeToRgbBlink1(device *Device, fadeTime time.Duration, red, green, blue, normal int) (bytesWritten int) {
+func fadeToRgbBlink1(device *Device, fadeTime time.Duration, red, green, blue, normal uint8) (bytesWritten int) {
 	bytesWritten = libusb.SendBlink1Command(device.Device, toMs(fadeTime), red, blue, green, normal)
 	return
 }
